requirements.txt
python-telegram-bot==20.1
python-telegram-bot[socks]==20.1
pymysql
sqlalchemy
emoji
itsdangerous
pyinstaller


